# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * extraGO — Workforce Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationCategory } from './notificationCategory';
import type { NotificationPriority } from './notificationPriority';

export interface Notification {
  id: number;
  userId: number;
  type: string;
  category?: NotificationCategory;
  priority?: NotificationPriority;
  title?: string;
  message?: string;
  isRead?: boolean;
  createdAt?: string;
  /** @nullable */
  link?: string | null;
}