# lib/api-zod/src/generated/types/notificationListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * extraGO — Workforce Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Notification } from './notification';

export interface NotificationListResponse {
  items: Notification[];
  total: number;
  unreadCount: number;
  page: number;
  limit: number;
  hasMore: boolean;
}